Updates made to rendering of app.
Extract the MenuItems into a separate component: This helps keep the code more modular and easier to read and maintain.

Using the Link component from react-router-dom for navigation instead of a Button with an href.
This is a more idiomatic way of navigating in a React app, and it allows React Router to handle
navigation without a full page refresh.

Using an array of objects to define your pages.
This could be helpful in case you want to add more properties to a page in the future.

Improve accessibility by adding a description to the IconButton.